Reto 4

+++ Reto_4/MMOpenMPI.c
@@ -13,7 +13,7 @@ void writeTime(double tiempo, int tam, int wnodos){
     fclose(f);
 }
 
-void printMat(int* mat, int n){
+void printMat(double* mat, int n){
     for(int i = 0; i < n; i++)
     {
         for(int j = 0; j < n; j++)
